requirements.txt
sqlalchemy==2.0.30
pymysql==1.1.0
python-dotenv==1.0.1